Adding graphic interface
Added interface to choose dataset and decide whether to rebuild the model, another one for setting learning parameters (just epochs and batch size for now) and another one for choosing which models to load. Removed functions for model loading that became unnecessary. Planning to make both English and Ukrainian versions with a separate root to choose the language. Another plan is to create a root for choosing whether to use a newly built model for prediction right away.